add snapshots for tokens package
