+++ packages/gestalt-design-tokens/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {},
   "devDependencies": {
+    "decomment": "^0.9.5",
     "style-dictionary": "^3.9.2",
     "tinycolor2": "1.6.0"
   },
