+++ scripts/test.sh
@@ -15,6 +15,9 @@ yarn prettier --check .
 echo "stylelint"
 yarn stylelint "**/*.css"
 
+echo "Tokens:Output Tests"
+yarn run tokens:test
+
 echo "jest"
 yarn jest --coverage
 
@@ -33,8 +36,7 @@ yarn run css:validate
 echo "Shell check"
 yarn run lint:sh
 
-echo "Tokens:Output Tests"
-yarn run tokens:test
+
 
 if [ -z "${GITHUB_ACTIONS:-}" ] && git diff HEAD master --name-only | grep -q packages/gestalt; then
   echo "Found changes to Components"
